+++ src/extension.ts
@@ -1,53 +1,49 @@
-'use strict'
-
+import debounce from 'lodash.debounce';
 import {
-    ExtensionContext,
+    commands,
     languages,
     window,
-    workspace
-} from 'vscode'
-import LinkProvider from './providers/linkProvider'
-import * as util    from './util'
+    workspace,
+} from 'vscode';
+import LinkProvider from './providers/linkProvider';
+import * as util from './util';
 
-let providers  = []
-const debounce = require('lodash.debounce')
+let providers = [];
 
-export function activate(context: ExtensionContext) {
-    util.readConfig()
+export function activate({ subscriptions }) {
+    util.readConfig();
 
     // config
     workspace.onDidChangeConfiguration((e) => {
         if (e.affectsConfiguration(util.PACKAGE_NAME)) {
-            util.readConfig()
+            util.readConfig();
         }
-    })
+    });
 
     // links
-    initProviders()
+    initProviders();
     window.onDidChangeActiveTextEditor(async (e) => {
-        await clearAll()
-        initProviders()
-    })
+        await clearAll();
+        initProviders();
+    });
 
     // scroll
-    util.scrollToText()
+    subscriptions.push(commands.registerCommand(util.CMND_NAME, util.scrollToText));
 }
 
-const initProviders = debounce(function() {
-    providers.push(languages.registerDocumentLinkProvider(['php', 'blade'], new LinkProvider()))
-}, 250)
+const initProviders = debounce(() => {
+    providers.push(languages.registerDocumentLinkProvider(['php', 'blade'], new LinkProvider()));
+}, 250);
 
 function clearAll() {
     return new Promise((res, rej) => {
-        providers.map((e) => e.dispose())
-        providers = []
+        providers.map((e) => e.dispose());
+        providers = [];
 
-        setTimeout(() => {
-            return res(true)
-        }, 500)
-    })
+        setTimeout(() => res(true), 500);
+    });
 }
 
 export function deactivate() {
-    clearAll()
+    clearAll();
 }