+++ CHANGELOG.md
@@ -13,3 +13,11 @@
 ## 1.0.3
 
 - Change icon
+
+## 2.0.0
+
+- fix link popup not being clickable
+- use a cmnd instead of the uri handler
+- use symbol provider to navigate to the correct config key
+- correct the link match regex
+- show output panel when phpcommand fails